feat: Add deposit functionality; implement DepositDto, DepositSchema, DepositUsecase, and corresponding tests; add ListCategoriesResponse and ListCategoriesUsecase for category management

+++ src/App/DTO/DepositDto.php
@@ -6,7 +6,7 @@ readonly class DepositDto
 {
     public function __construct(
         public string $accountId,
-        public float $amount,
+        public int $amount,
         public string $categoryId,
         public ?string $description = null,
     ) {
